requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
google-generativeai>=0.3.2
langchain-community>=0.0.21
streamlit>=1.30.0
python-dotenv>=1.0.0
